Add file `explcheck-config.lua` with default configuration of explcheck
You may place a file named `explcheck-config.lua` with your own
configuration in your repository to control the behavior of explcheck.

Note that the configuration options are provisional and may be changed
or removed before version 1.0.0. Furthermore, support for configuration
YAML files that will allow you to specify different configuration for
different .tex files is envisioned for a future release and will be the
recommended way to configure explcheck.